requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu124
torch>=2.4.0+cu124
torchaudio>=2.4.0+cu124
torchvision>=0.19.0+cu124
transformers>=4.41.1
accelerate>=0.28.0
quanto
optimum-quanto  # bitsandbytes replacement, seems better?
sentence-transformers
einops
einops_exts
open_clip_torch>=2.24.0
treelib
nltk>=3.9
# multi-rake
yake
symspellpy
rapidfuzz
streamlit
streamlit-pdf-viewer
opencv-python
pdf2image
pytesseract
pdfplumber>=0.11.3
pdfminer.six>=20231228  # fixes infinite loop from unstructured[all-docs] of PDFMiner Read In?
unstructured[all-docs]>=0.15.5
llama-index-core
llama-index-embeddings-huggingface
llama-index-vector-stores-qdrant
llama-index-retrievers-bm25
llama-index-llms-huggingface
llama-index-llms-groq
llama-index-question-gen-openai  # required for subquestionqueryengine
llama-index-multi-modal-llms-openai